+++ examples/randwalk/Makefile
@@ -1,0 +1,11 @@
+#
+#	Makefile for this SNAP example
+#	- modify Makefile.ex when creating a new SNAP example
+#
+#	implements:
+#		all (default), clean
+#
+
+include ../../Makefile.config
+include Makefile.ex
+include ../Makefile.exmain
