+++ examples/randwalk/Makefile.ex
@@ -1,0 +1,8 @@
+#
+#	configuration variables for the example
+
+## Main application file
+MAIN = randwalk
+DEPH = 
+DEPCPP = 
+
