+++ examples/randwalk/stdafx.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "targetver.h"
+
+#include "Snap.h"
+
